scripts/python/requirements.txt
# Layout detection dependencies
layoutparser==0.3.4
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0

# Detectron2 must be installed separately:
# pip install 'git+https://github.com/facebookresearch/detectron2.git'

